requirements.txt
eel
openai
python-dotenv
